return Object instead of Integer for Pyro predictions
Replace imp with importlib for Python version > 3 in pyroAdapter.py

+++ cemlComponents/models/PythonModelsIntegrator/src/main/java/eu/linksmart/services/event/ceml/models/PythonPyroBase.java
@@ -1,39 +1,41 @@
 package eu.linksmart.services.event.ceml.models;
 
-import com.fasterxml.jackson.core.Version;
-import com.fasterxml.jackson.databind.module.SimpleModule;
-import com.fasterxml.jackson.databind.type.TypeFactory;
-import eu.linksmart.api.event.ceml.data.ClassesDescriptor;
 import eu.linksmart.api.event.ceml.evaluation.TargetRequest;
-import eu.linksmart.api.event.ceml.model.Model;
-import eu.linksmart.api.event.ceml.model.ModelDeserializer;
 import eu.linksmart.api.event.ceml.prediction.Prediction;
 import eu.linksmart.api.event.ceml.prediction.PredictionInstance;
-import eu.linksmart.api.event.exceptions.*;
-import eu.linksmart.services.event.ceml.evaluation.evaluators.DoubleTumbleWindowEvaluator;
+import eu.linksmart.api.event.exceptions.InternalException;
+import eu.linksmart.api.event.exceptions.StatementException;
+import eu.linksmart.api.event.exceptions.TraceableException;
+import eu.linksmart.api.event.exceptions.UntraceableException;
 import eu.linksmart.services.event.intern.Const;
 import eu.linksmart.services.event.intern.SharedSettings;
-
-import java.io.*;
-import java.util.*;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
-
 import eu.linksmart.services.utils.configuration.Configurator;
 import eu.linksmart.services.utils.function.Utils;
 import io.swagger.client.ApiClient;
 import io.swagger.client.api.ScApi;
-import net.razorvine.pyro.*;
+import net.razorvine.pyro.NameServerProxy;
+import net.razorvine.pyro.PyroException;
+import net.razorvine.pyro.PyroProxy;
+import net.razorvine.pyro.PyroURI;
 import org.apache.commons.io.FileUtils;
 import org.apache.http.client.fluent.Request;
 import org.apache.http.entity.ContentType;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import java.io.File;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+
 /**
  * Created by José Ángel Carvajal on 06.02.2018 a researcher of Fraunhofer FIT.
  */
-public class PythonPyroBase<T> extends ClassifierModel<T,Integer,PyroProxy>{
+public class PythonPyroBase<T> extends ClassifierModel<T, Object, PyroProxy> {
     public transient static Logger loggerService = LogManager.getLogger(PythonPyroBase.class);
     protected Process proc;
     static protected Process orchProc;
@@ -42,28 +44,27 @@ public class PythonPyroBase<T> extends ClassifierModel<T,Integer,PyroProxy>{
     protected int counter = 0;
     protected final static String pyroAdapterFilename = "pyroAdapter.py";
     private static final Configurator conf = Configurator.getDefaultConfig();
-    protected final static String pythonPath ;
+    protected final static String pythonPath;
+
     static {
         String path = "";
         try {
             path = conf.getString(Const.PYTHON_PATH);
-        }catch (Exception e){
-            loggerService.error(e.getMessage(),e);
+        } catch (Exception e) {
+            loggerService.error(e.getMessage(), e);
         }
         path.trim();
-        if(!path.isEmpty() && path.charAt(path.length()-1) != File.separator.charAt(0))
-            path=path+File.separator;
+        if (!path.isEmpty() && path.charAt(path.length() - 1) != File.separator.charAt(0))
+            path = path + File.separator;
         pythonPath = path;
     }
+
     protected Request orchestrator = null;
     protected Object baseModel = null;
 
-
     public PythonPyroBase(List<TargetRequest> targets, Map<String, Object> parameters, Object learner) {
         super(targets, parameters, null);
-        baseModel= learner;
-
-
+        baseModel = learner;
     }
 
     @Override
@@ -71,59 +72,58 @@ public class PythonPyroBase<T> extends ClassifierModel<T,Integer,PyroProxy>{
 
         ScApi SCclient;
 
-
         try {
-            LinkedHashMap backend = (LinkedHashMap)parameters.get("Backend");
+            LinkedHashMap backend = (LinkedHashMap) parameters.get("Backend");
 
-            if((boolean)backend.getOrDefault("Lookup",false)) {
+            if ((boolean) backend.getOrDefault("Lookup", false)) {
                 // Lookup a running agent
                 loggerService.info("Looking up '{}' in name server '{}'", backend.get("RegisteredName"), backend.get("NameServer"));
-                NameServerProxy ns = NameServerProxy.locateNS((String)backend.get("NameServer"));
-                learner = new PyroProxy(ns.lookup((String)backend.get("RegisteredName")));
+                NameServerProxy ns = NameServerProxy.locateNS((String) backend.get("NameServer"));
+                learner = new PyroProxy(ns.lookup((String) backend.get("RegisteredName")));
                 ns.close();
-            }if(backend.get("Orchestrator")!=null) {
-                String lernerUri, orchURL;
-                if (backend.get("OrchestratorName")!=null) {
-                    // search for existing orchestrator
-                    loggerService.info("Looking up '{}' in name server '{}'", backend.get("OrchestratorName"), backend.get("NameServer"));
-                    if (Utils.isRestAvailable(conf.getString(eu.linksmart.services.utils.constants.Const.LINKSMART_SERVICE_CATALOG_ENDPOINT))) {
-                        ApiClient apiClient = new ApiClient();
-                        apiClient.setBasePath(conf.getString(eu.linksmart.services.utils.constants.Const.LINKSMART_SERVICE_CATALOG_ENDPOINT));
-                        SCclient = new ScApi(apiClient);
-                        orchURL = SCclient.idGet(String.valueOf(backend.containsValue("OrchestratorName"))).getApis().get("HTTP");
+
+                if (backend.get("Orchestrator") != null) {
+                    String lernerUri, orchURL;
+                    if (backend.get("OrchestratorName") != null) {
+                        // search for existing orchestrator
+                        loggerService.info("Looking up '{}' in name server '{}'", backend.get("OrchestratorName"), backend.get("NameServer"));
+                        if (Utils.isRestAvailable(conf.getString(eu.linksmart.services.utils.constants.Const.LINKSMART_SERVICE_CATALOG_ENDPOINT))) {
+                            ApiClient apiClient = new ApiClient();
+                            apiClient.setBasePath(conf.getString(eu.linksmart.services.utils.constants.Const.LINKSMART_SERVICE_CATALOG_ENDPOINT));
+                            SCclient = new ScApi(apiClient);
+                            orchURL = SCclient.idGet(String.valueOf(backend.containsValue("OrchestratorName"))).getApis().get("HTTP");
+                        } else {
+                            throw new StatementException(this.getName(), "Bad Request", "The orchestration was unable to be obtain from Service Catalog");
+                        }
+                    } else if (backend.get("OrchestratorURL") != null) {
+                        // search for existing orchestrator
+                        loggerService.info("Looking up '{}' in name server '{}'", backend.get("OrchestratorURL"));
+                        orchURL = String.valueOf(backend.get("OrchestratorURL")) + "/pyro/";
+                        // orchestrator = new PyroProxy(ns.lookup((String)backend.get("OrchestratorName")));
+
                     } else {
-                        throw new StatementException(this.getName(), "Bad Request", "The orchestration was unable to be obtain from Service Catalog");
+                        throw new StatementException(this.getName(), "Bad Request", "The orchestration option was given but the orchestrator was not provided");
                     }
 
-                } else if (backend.get("OrchestratorURL")!=null) {
-                    // search for existing orchestrator
-                    loggerService.info("Looking up '{}' in name server '{}'", backend.get("OrchestratorURL"));
-                    orchURL = String.valueOf(backend.get("OrchestratorURL")) + "/pyro/";
-                    // orchestrator = new PyroProxy(ns.lookup((String)backend.get("OrchestratorName")));
+                    Request request = Request.Post(orchURL).bodyString(SharedSettings.getSerializer().toString(backend.get("Orchestrator")), ContentType.APPLICATION_JSON);
+                    Map response = SharedSettings.getDeserializer().parse(request.execute().returnContent().asString(), Map.class);
+                    lernerUri = response.get("uri").toString();
+                    learner = new PyroProxy(new PyroURI(lernerUri));
 
-                } else {
-                    throw new StatementException(this.getName(), "Bad Request", "The orchestration option was given but the orchestrator was not provided");
+                    // loggerService.info("Orchestrator: {}:{} {}", orchestrator.hostname, orchestrator.port, orchestrator.pyroHandshake);
+
                 }
-
-                Request request = Request.Post(orchURL).bodyString(SharedSettings.getSerializer().toString(backend.get("Orchestrator")), ContentType.APPLICATION_JSON);
-                Map response = SharedSettings.getDeserializer().parse(request.execute().returnContent().asString(), Map.class);
-                lernerUri = response.get("uri").toString();
-                learner = new PyroProxy(new PyroURI(lernerUri));
-
-                // loggerService.info("Orchestrator: {}:{} {}", orchestrator.hostname, orchestrator.port, orchestrator.pyroHandshake);
-
-            }else {
-                //  create temp python script for the pythonAdapter
+            } else {
+                //create temp python script for the pythonAdapter
                 pyroAdapter = copyPyScripts(pyroAdapterFilename);
-                // run an adapter form scratch
-               learner = constructProxy(pyroAdapter,proc,backend.getOrDefault("ModuleName","").toString());
+                //run an adapter form scratch
+                learner = constructProxy(pyroAdapter, proc, backend.getOrDefault("ModuleName", "").toString());
             }
 
             loggerService.info("Learner: {}:{} {}", learner.hostname, learner.port, learner.pyroHandshake);
-            if(baseModel != null)
+            if (baseModel != null)
                 learner.call("importModel", baseModel);
             learner.call("build", parameters.get("Classifier"));
-
         } catch (PyroException e) {
             loggerService.error(e._pyroTraceback);
             throw new InternalException(this.getName(), "Pyro", e);
@@ -136,29 +136,28 @@ public class PythonPyroBase<T> extends ClassifierModel<T,Integer,PyroProxy>{
     }
 
     private File copyPyScripts(String pyScriptFilename) throws IOException {
-        File file = new File(System.getProperty("java.io.tmpdir")+File.separator+pyScriptFilename);
+        File file = new File(System.getProperty("java.io.tmpdir") + File.separator + pyScriptFilename);
         file.deleteOnExit();
         FileUtils.copyURLToFile(this.getClass().getClassLoader().getResource(pyScriptFilename), file);
         return file;
     }
-    private PyroProxy constructProxy(File file, Process proc,String moduleName) throws IOException, InternalException {
 
-       // LinkedHashMap backend = (LinkedHashMap)parameters.get("Backend");
+    private PyroProxy constructProxy(File file, Process proc, String moduleName) throws IOException, InternalException {
+
+        // LinkedHashMap backend = (LinkedHashMap)parameters.get("Backend");
 
         loggerService.info("Saved script to {}", file.getAbsolutePath());
 
         // Path to script is passed as parameter
-        String[] cmd = {pythonPath+"python", "-u", file.getAbsolutePath(),
-                "--bname="+ moduleName,
-                "--bpath="+ file.getAbsolutePath(),
-                "--orch="+ ("Orchestrator".equals(moduleName))};
+        String[] cmd = {pythonPath + "python", "-u", file.getAbsolutePath(),
+                "--bname=" + moduleName,
+                "--bpath=" + file.getAbsolutePath(),
+                "--orch=" + ("Orchestrator".equals(moduleName))};
 
-        String agentPyroURI = Utils.runGetLastOutput(cmd,moduleName,loggerService);
+        String agentPyroURI = Utils.runGetLastOutput(cmd, moduleName, loggerService);
 
         return new PyroProxy(new PyroURI(agentPyroURI));
     }
-
-
 
     @Override
     public synchronized void learn(T input) throws TraceableException, UntraceableException {
@@ -173,9 +172,9 @@ public class PythonPyroBase<T> extends ClassifierModel<T,Integer,PyroProxy>{
     }
 
     @Override
-    public synchronized Prediction<Integer> predict(T input) throws TraceableException, UntraceableException {
+    public synchronized Prediction<Object> predict(T input) throws TraceableException, UntraceableException {
         try {
-            return toPrediction(input, (Integer) learner.call("predict", input));
+            return toPrediction(input, (Object) learner.call("predict", input));
         } catch (PyroException e) {
             loggerService.error(e._pyroTraceback);
             throw new InternalException(this.getName(), "Pyro", e);
@@ -209,7 +208,7 @@ public class PythonPyroBase<T> extends ClassifierModel<T,Integer,PyroProxy>{
     }
 
     @Override
-    public synchronized void batchLearn(List<T> input) throws TraceableException, UntraceableException{
+    public synchronized void batchLearn(List<T> input) throws TraceableException, UntraceableException {
         try {
             learner.call("batchLearn", input);
         } catch (PyroException e) {
@@ -221,11 +220,11 @@ public class PythonPyroBase<T> extends ClassifierModel<T,Integer,PyroProxy>{
     }
 
     @Override
-    public synchronized List<Prediction<Integer>> batchPredict(List<T> input) throws TraceableException, UntraceableException{
+    public synchronized List<Prediction<Object>> batchPredict(List<T> input) throws TraceableException, UntraceableException {
         counter += (int) parameters.get("RetrainEvery");
 
         try {
-            List<Integer> res = (List<Integer>) learner.call("batchPredict", input);
+            List<Object> res = (List<Object>) learner.call("batchPredict", input);
 
             if (input.size() != res.size())
                 throw new InternalException(this.getName(), "Pyro", "Batch predictions are not the same size as inputs.");
@@ -243,15 +242,15 @@ public class PythonPyroBase<T> extends ClassifierModel<T,Integer,PyroProxy>{
     public synchronized void destroy() throws Exception {
         learner.call("destroy");
         learner.close();
-        if(proc!=null)
+        if (proc != null)
             proc.destroy();
-        if(pyroAdapter.exists())
+        if (pyroAdapter != null && pyroAdapter.exists())
             pyroAdapter.delete();
         super.destroy();
     }
 
-    // Convert integer prediction to Prediction<Integer>
-    private Prediction<Integer> toPrediction(Object input, Integer response) {
-        return new PredictionInstance<>(response,input, SharedSettings.getId()+":"+this.getName(),new ArrayList<>(evaluator.getEvaluationAlgorithms().values()));
+    // Convert Object prediction to Prediction<Object>
+    private Prediction<Object> toPrediction(Object input, Object response) {
+        return new PredictionInstance<>(response, input, SharedSettings.getId() + ":" + this.getName(), new ArrayList<>(evaluator.getEvaluationAlgorithms().values()));
     }
 }
